+++ common++/Bank/Bankauftrag.cc
@@ -1,4 +1,4 @@
-/* $Id: Bankauftrag.cc,v 1.11 2003/05/13 17:32:42 jacek Exp $ */
+/* $Id: Bankauftrag.cc,v 1.12 2003/11/07 11:12:50 jacek Exp $ */
 /*  libcommonc++: ManuProC's main OO library
  *  Copyright (C) 1998-2000 Adolf Petig GmbH & Co. KG, written by Christof Petig
  *
@@ -182,7 +182,8 @@ void Bankauftrag::close(bool nodisk) throw(IOerror)
    fprintf(druckerpipe,   "\\pagestyle{empty}\n"
 	   "Disketten-Begleitzettel vom %s, Belegloser Datenträgeraustausch"
 	   "\\hfill ReferenzNr. %s\\\\\n"
-	   "\n",ManuProC::Datum(Kennziffer).c_str(),Formatiere(Kennziffer).c_str());
+	   "\n",ManuProC::Datum(Kennziffer).c_str(),
+			Formatiere((unsigned long)Kennziffer).c_str());
    
    fprintf(druckerpipe,"\\begin{tabular}{lr}\n");
    fprintf(druckerpipe,"\\multicolumn{2}{l}{Datei \\texttt{DTAUS1.TXT} für %s}\\\\\n",
@@ -197,7 +198,8 @@ void Bankauftrag::close(bool nodisk) throw(IOerror)
    "Summe  Datensätze C (\\euro{}):&%s\\\\\n",Formatiere((unsigned long long)Betragssumme,2).c_str());
    fprintf(druckerpipe,"Anzahl Datensätze C:&%d\\\\\n",anzahlDatensaetze);
    fprintf(druckerpipe,"Summe der Kontonummern:&%s\\\\\n",Formatiere((unsigned long long)Kontosumme).c_str());
-   fprintf(druckerpipe,"Summe der Bankleitzahlen:&%s\\\\[2cm]\n",Formatiere(BLZsumme).c_str());
+   fprintf(druckerpipe,"Summe der Bankleitzahlen:&%s\\\\[2cm]\n",
+		Formatiere((unsigned long long)BLZsumme).c_str());
    
    fprintf(druckerpipe,"Wuppertal, %s\\\\[2cm]\n",ManuProC::Datum(Kennziffer).c_str());
    fprintf(druckerpipe,"\\multicolumn{2}{l}{\\rule{10cm}{0.5pt}}\\\\\n"
